Moved formatChapterType to common trustees.ts
Also fixed tests.

Jira ticket: CAMS-640
